Legger mekking av riktig rest-format for sanity-begrunnelser inn i service
